# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: Students Week 2026 food ordering API
servers:
  - url: /api
...
paths:
...
  /vendors:
    get:
      operationId: getVendors
      tags: [vendors]
      summary: List vendors
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    patch:
      operationId: updateVendor
      tags: [vendors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /tables:
    get:
      operationId: getTables
      tags: [tables]
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventTable"
    post:
      operationId: createTable
      tags: [tables]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventTable"
  /tables/{id}:
    patch:
      operationId: updateTable
      tags: [tables]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventTable"
    delete:
      operationId: deleteTable
      tags: [tables]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /vendor-dashboard/{vendorId}/stats:
    get:
      operationId: getVendorDashboardStats
      tags: [vendor]
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDashboardStats"
components:
  schemas:
...
    Vendor:
      type: object
      required: [id, name, isActive, maxPlates, orderCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        maxPlates:
          type: integer
        vendorPin:
          type: ["string", "null"]
        orderCount:
          type: integer
        createdAt:
          type: string
    VendorInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        imageUrl:
          type: string
        isActive:
          type: boolean
        maxPlates:
          type: integer
          minimum: 0
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        maxPlates:
          type: integer
          minimum: 0
        vendorPin:
          type: ["string", "null"]
...
    EventTable:
      type: object
      required: [id, tableNumber, isActive, createdAt]
      properties:
        id:
          type: integer
        tableNumber:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    TableInput:
      type: object
      required: [tableNumber]
      properties:
        tableNumber:
          type: string
          minLength: 1
        isActive:
          type: boolean
    TableUpdate:
      type: object
      properties:
        tableNumber:
          type: string
          minLength: 1
        isActive:
          type: boolean
...
    VendorDashboardStats:
      type: object
      required: [totalOrders, pendingOrders, servedOrders, remainingPlates]
      properties:
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        servedOrders:
          type: integer
        remainingPlates:
          type: integer
        popularMainDishes:
          type: array
          items:
            type: object
            required: [name, count]
            properties:
              name:
                type: string
              count:
                type: integer
        popularMainDishTypes:
          type: array
          items:
            type: object
            required: [name, count]
            properties:
              name:
                type: string
              count:
                type: integer
        popularSides:
          type: array
          items:
            type: object
            required: [name, count]
            properties:
              name:
                type: string
              count:
                type: integer
        popularProteins:
          type: array
          items:
            type: object
            required: [name, count]
            properties:
              name:
                type: string
              count:
                type: integer